fix bugfix

+++ be/src/exec/partition/chunks_partitioner.h
@@ -75,7 +75,7 @@ public:
     }
 
     // Number of partitions
-    size_t num_partitions() const { return _hash_map_variant.size(); }
+    int32_t num_partitions() const { return _hash_map_variant.size(); }
 
     bool is_passthrough() const { return _is_passthrough; }
